tests: dsd: fix va_args guint64 type promotion on wasm32
On wasm32, int is not promoted to guint64 in variadic calls. Passing
the literal 0x3 (int) where GST_TYPE_BITMASK expects guint64 reads
garbage from the stack, causing channel position lookup to fail.

Cast the bitmask value to (guint64) explicitly.

+++ subprojects/gst-plugins-base/tests/check/libs/dsd.c
@@ -291,7 +291,7 @@ GST_START_TEST (test_dsd_info_from_caps)
       "rate", G_TYPE_INT, GST_DSD_MAKE_DSD_RATE_44x (128),
       "channels", G_TYPE_INT, 2,
       "layout", G_TYPE_STRING, "non-interleaved",
-      "channel-mask", GST_TYPE_BITMASK, 0x3, NULL);
+      "channel-mask", GST_TYPE_BITMASK, (guint64) 0x3, NULL);
   fail_unless (gst_dsd_info_from_caps (&info, full_caps));
   fail_unless_equals_int (info.format, GST_DSD_FORMAT_U16LE);
   fail_unless_equals_int (info.rate, GST_DSD_MAKE_DSD_RATE_44x (128));
